# files-api-spec.yaml
openapi: 3.0.0
info:
  description: 'The HuBMAP Files API is a standard RESTful web service to manage files associated with project entities.'
  version: 0.1.0
  title: HuBMAP Files API
  contact:
    name: HuBMAP Help Desk
    email: help@hubmapconsortium.org
  license:
    name: MIT License
    url: 'https://github.com/hubmapconsortium/files-api/blob/main/LICENSE'
servers:
  - url: 'https://files.api.hubmapconsortium.org'
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: Globus bearer token to authorize requests. Have to use apiKey to pass AWS API Gateway import validation
  schemas:
paths:
  '/entities/{id}/files':
    get:
      summary: Get the information about files attached to an entity, given an entity identifier (e.g. HBM123.ABCD.456 or UUID.)
      parameters:
        - name: id
          in: path
          description: Either an HuBMAP ID (e.g. HBM123.ABCD.456) or UUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A JSON array containing an object for each file attached to the entity.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    file_uuid:
                      type: string
                      description: The uuid of the file
                    path:
                      type: string
                      description: The local file system path of the file relative to its Dataset directory, including the file name
                    checksum:
                      type: string
                      description: The checksum of the file
                    size:
                      type: integer
                      description: Integer size of the file in bytes
                    base_dir:
                      type: string
                      enum:
                        - INGEST_PORTAL_UPLOAD
                        - DATA_UPLOAD
                      description: One of DATA_UPLOAD|INGEST_PORTAL_UPLOAD. INGEST_PORTAL_UPLOAD means the file was uploaded into the space for file uploads from the Ingest UI.  DATA_UPLOAD means the file was upload into the upload space for datasets, usually via Globus
        '303':
          description: The response is too large for the endpoint to return, and must be retrieved from an AWS S3 bucket within one hour, at the URL in response. The bucket will contain the same JSON content as described for an HTTP 200 response.'
        '401':
          description: The user's token has expired or the user did not supply a valid token
        '403':
          description: The user is not authorized to use this method
        '404':
          description: The target id could not be found
        '500':
          description: An unexpected error occurred
  '/datasets/{id}/reindex':
    put:
      summary: Create a file info JSON document for each File in the public Dataset of the UUID provided, and put it in the Elasticsearch hm_files indices. Public Datasets are published and do not contain genetic information.
      parameters:
        - name: id
          in: path
          description: A Dataset 32-character UUID or HuBMAP ID (e.g. HBM123.ABCD.456)
          required: true
          schema:
            type: string
      responses:
        '202':
          description: A response acknowledging the request to load the Dataset, including the Dataset UUID and the index name.